Extract Workspace into separate file.
